+++ src/Graph/Nodes/Outcome.php
@@ -16,9 +16,15 @@ use Stringable;
 interface Outcome extends Decision, BinaryNode, Stringable
 {
     /**
-     * Return the outcome of the decision, depends on the actual class.
+     * <<<<<<< HEAD
+     * Return the outcome of the decision.
+     *
+     * @return string|int|float
+     *                          =======
+     *                          Return the outcome of the decision, depends on the actual class.
      *
      * @return mixed
+     *               >>>>>>> 2.1
      */
     public function outcome();
 }