+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - Added Probabilistic and Top K Accuracy
     - Added Brier Score Probabilistic Metric
     - Export Decision Tree-based models in Graphviz "dot" format
+    - Added Graphviz helper class
     - Graph subsystem memory and storage optimizations
     
 - 2.0.2